Prevent VSCode from opening when remounting
